requirements.txt
phidata
python-dotenv
duckduckgo-search
groq
requests